tests/requirements.txt
coverage>=5.1
coverage-pyver-pragma>=0.0.6
domdf-python-tools[testing]>=2.0.1
iniconfig!=1.1.0,>=1.0.1
pytest-cov>=2.8.1
pytest-randomly>=3.3.1
pytest-timeout>=1.4.2
